requirements.txt
Flask==2.2.2
gunicorn==20.1.0
mysql-connector-python==8.0.33
python-dotenv==0.21.0